updated the player controller, input manager and created the player camera script
updated the player controller to have an empty attack function, used in testing

updated the input manager so that it now knows about the camera movement inputs and can respond accordingly

made the player camera script, this script is used to move the camera to where the player would like to look, this script needs to be on the main camera, with the player object in its target  gameobject variable slot. this camera functions using the left and right mouse buttons, can be changed at a later date if need be.