Added Health Perc to the Player
Player.cs has a health percentage added to it, Currently in Update for testing, with a Debug log to watch in the editor, the players currentHealth vairable is public for testing, needs to be changed back to protected

No longer using PlayerObject class, it is fully commented out, but being left in for Steve.

Fixed the Calculation in the Character base for a health percentage, dividing maxHealth by current, giving a higher number hte less health the player had, as well as not giving a numner out of 100.